+++ R/funky_heatmap.R
@@ -40,13 +40,10 @@
 #' * `hjust`: Horizontal alignment of the bar, must be between \[0,1\]
 #'     (only for `geom = "bar"`).
 #'
-#' * `hjust`: Horizontal alignment of the label, must be between \[0,1\]
-#'     (only for `geom = "text"`).
-#'
 #' * `vjust`: Vertical alignment of the label, must be between \[0,1\]
 #'     (only for `geom = "text"`).
 #'
-#' * `size`: Size of the label, must be between \[0,1\]
+#' * `size`: Size of the label, must be a numeric value
 #'     (only for `geom = "text"`).
 #'
 #' * `label`: Which column to use as a label (only for `geom = "text"`).
@@ -96,7 +93,29 @@
 #'
 #' * `numerical`: `"Greys"`, `"Blues"`, `"Reds"`, `"YlOrBr"`, `"Greens"`
 #' * `categorical`: `"Set3"`, `"Set1"`, `"Set2"`, `"Dark2"`
+#' 
+#' @param legends A list of legends to add to the plot. Each entry in
+#' `column_info$legend` should have a corresponding entry in this object.
+#' Each entry should be a list with the following names:
+#'  * `palette` (`character`): The palette to use for the legend. Must be
+#'   a value in `palettes`.
+#'  * `geom` (`character`): The geom of the legend. Must be one of:
+#'    `"funkyrect"`, `"circle"`, `"rect"`, `"bar"`, `"pie"`, `"text"`, `"image"`.
+#'  * `title` (`character`, optional): The title of the legend. Defaults
+#'    to the palette name.
+#'  * `enabled` (`logical`, optional): Whether or not to add the legend.
+#'    Defaults to `TRUE`.
+#'  * `labels` (`character`, optional): The labels to use for the legend.
+#'    The defaults depend on the selected geom.
+#'  * `size` (`numeric`, optional): The size of the listed geoms.
+#'    The defaults depend on the selected geom.
+#'  * `color` (`character`, optional): The color of the listed geoms.
+#'    The defaults depend on the selected geom.
+#'  * `value` (optional): Used as values for the text and image geoms.
 #'
+#' @param position_args Sets parameters that affect positioning within a
+#' plot, such as row and column dimensions, annotation details, and the
+#' expansion directions of the plot. See `position_arguments()` for more information.
 #'
 #' @param scale_column Whether or not to apply min-max scaling to each
 #' numerical column.
@@ -104,15 +123,13 @@
 #' @param add_abc Whether or not to add subfigure labels to the different
 #' columns groups.
 #'
-#' @param col_annot_offset How much the column annotation will be offset by.
-#' @param col_annot_angle The angle of the column annotation labels.
-#' @param removed_entries Which methods to not show in the rows. Missing methods
-#' are replaced by a "Not shown" label.
-#'
-#' @param expand A list of directions to expand the plot in.
+#' @param col_annot_offset DEPRECATED: use `position_args = position_arguments(col_annot_offset = ...)` instead.
+#' @param col_annot_angle DEPRECATED: use `position_args = position_arguments(col_annot_angle = ...)` instead.
+#' @param expand DEPRECATED: use `position_args = position_arguments(expand_* = ...)` instead.
 #'
 #' @importFrom ggforce geom_arc_bar geom_circle geom_arc
 #' @importFrom cowplot theme_nothing
+#' @importFrom patchwork wrap_plots plot_spacer
 #'
 #' @returns A ggplot. `.$width` and `.$height` are suggested dimensions for
 #' storing the plot with [ggsave()].
@@ -128,18 +145,20 @@
 #'
 #' funky_heatmap(data)
 funky_heatmap <- function(
-    data,
-    column_info = NULL,
-    row_info = NULL,
-    column_groups = NULL,
-    row_groups = NULL,
-    palettes = NULL,
-    scale_column = TRUE,
-    add_abc = TRUE,
-    col_annot_offset = 3,
-    col_annot_angle = 30,
-    removed_entries = NULL,
-    expand = c(xmin = 0, xmax = 2, ymin = 0, ymax = 0)) {
+  data,
+  column_info = NULL,
+  row_info = NULL,
+  column_groups = NULL,
+  row_groups = NULL,
+  palettes = NULL,
+  legends = NULL,
+  position_args = position_arguments(),
+  scale_column = TRUE,
+  add_abc = TRUE,
+  col_annot_offset,
+  col_annot_angle,
+  expand
+) {
   # validate input objects
   data <- verify_data(data)
   column_info <- verify_column_info(column_info, data)
@@ -147,6 +166,23 @@ funky_heatmap <- function(
   column_groups <- verify_column_groups(column_groups, column_info)
   row_groups <- verify_row_groups(row_groups, row_info)
   palettes <- verify_palettes(palettes, column_info, data)
+  legends <- verify_legends(legends, palettes, column_info, data)
+
+  # check deprecated arguments
+  if (!missing(col_annot_offset)) {
+    warning("Argument `col_annot_offset` is deprecated. Use `position_arguments(col_annot_offset = ...)` instead.")
+    position_args$col_annot_offset <- col_annot_offset
+  }
+  if (!missing(col_annot_angle)) {
+    warning("Argument `col_annot_angle` is deprecated. Use `position_arguments(col_annot_angle = ...)` instead.")
+    position_args$col_annot_angle <- col_annot_angle
+  }
+  if (!missing(expand)) {
+    warning("Argument `expand` is deprecated. Use `position_arguments(expand_* = ...)` instead.")
+    for (name in names(expand)) {
+      position_args[[paste0("expand_", name)]] <- expand[[name]]
+    }
+  }
   # todo: add column groups to verify_palettes
 
   geom_positions <- calculate_geom_positions(
@@ -156,15 +192,71 @@ funky_heatmap <- function(
     column_groups,
     row_groups,
     palettes,
+    position_args,
     scale_column,
-    add_abc,
-    col_annot_offset,
-    col_annot_angle,
-    removed_entries
+    add_abc
   )
 
-  compose_ggplot(
+  main_plot <- compose_ggplot(
     geom_positions,
-    expand
+    position_args
   )
+
+  # start plotting legends
+  geom_legend_funs <- list(
+    funkyrect = create_funkyrect_legend,
+    circle = create_circle_legend,
+    rect = create_rect_legend,
+    pie = create_pie_legend,
+    text = create_text_legend
+    # image = create_image_legend
+    # todo: add text legend
+    # todo: add bar legend
+  )
+  legend_plots <- list()
+  for (legend in legends) {
+    if (legend$enabled) {
+      legend_fun <- geom_legend_funs[[legend$geom]]
+      legend_args <- legend
+      legend_args$geom <- NULL
+      legend_args$enabled <- NULL
+      if (!is.null(legend$palette)) {
+        legend_args$palette <- palettes[[legend$palette]]
+      }
+      legend_args$position_args <- position_args
+      legend_plot <- do.call(legend_fun, legend_args)
+      legend_plots <- c(legend_plots, list(legend_plot))
+    }
+  }
+
+  if (length(legend_plots) == 0) {
+    return(main_plot)
+  }
+
+  legend_widths <- map_dbl(legend_plots, ~ .x$width)
+  legend_heights <- map_dbl(legend_plots, ~ .x$height)
+  
+  heights <- main_plot$height
+  width <- main_plot$width
+
+  heights <- c(heights, .1, max(legend_heights))
+  width <- max(width, sum(legend_widths))
+
+  out <- patchwork::wrap_plots(
+    main_plot,
+    patchwork::plot_spacer(),
+    patchwork::wrap_plots(
+      legend_plots,
+      nrow = 1,
+      widths = legend_widths
+    ),
+    ncol = 1,
+    heights = heights
+  )
+
+  # TODO: fix this heuristic
+  out$width <- width
+  out$height <- sum(heights)
+
+  out
 }