+++ R/geom_rounded_rect.R
@@ -60,27 +60,30 @@ geom_rounded_rect <- function(
 GeomRoundedRect <- ggplot2::ggproto(
   "GeomRoundedRect", ggplot2::Geom,
   default_aes = ggplot2::aes(
-    colour = NA, fill = "grey35", size = 0.5, linetype = 1, alpha = NA,
+    colour = NA, fill = "grey35", linewidth = 0.5, linetype = 1, alpha = NA,
     radius = 0.5
   ),
   required_aes = c("xmin", "xmax", "ymin", "ymax", "radius"),
-  draw_panel = function(self, data, panel_params, coord) {
+
+  draw_panel = function(self, data, panel_params, coord, lineend = "butt", linejoin = "mitre") {
     coords <- coord$transform(data, panel_params)
 
     gl <- lapply(seq_along(coords$xmin), function(i) {
       grid::roundrectGrob(
-        coords$xmin[i], coords$ymax[i],
+        x = coords$xmin[i],
+        y = coords$ymax[i],
         width = (coords$xmax[i] - coords$xmin[i]),
-        height = (coords$ymax[i] - coords$ymin)[i],
+        height = (coords$ymax[i] - coords$ymin[i]),
         r = grid::unit(coords$radius[i], "native"),
         default.units = "native",
         just = c("left", "top"),
         gp = grid::gpar(
           col = coords$colour[i],
           fill = alpha(coords$fill[i], coords$alpha[i]),
-          lwd = coords$size[i] * .pt,
+          lwd = coords$linewidth[i] * .pt,
           lty = coords$linetype[i],
-          lineend = "butt"
+          linejoin = linejoin,
+          lineend = lineend
         )
       )
     })
@@ -90,5 +93,6 @@ GeomRoundedRect <- ggplot2::ggproto(
     grob$name <- grid::grobName(grob, "geom_rounded_rect")
     grob
   },
-  draw_key = ggplot2::draw_key_polygon
+  draw_key = ggplot2::draw_key_polygon,
+  rename_size = TRUE
 )